requirements.txt
appnope==0.1.0
APScheduler==3.6.3
attrs==19.3.0
awsebcli==3.18.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blessed==1.17.4
blinker==1.4
blis==0.4.1
boto3==1.12.46
botocore==1.15.46
bs4==0.0.1
cached-property==1.5.1
catalogue==1.0.0
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.9.2
cymem==2.0.3
decorator==4.4.2
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
en-core-web-sm==2.2.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
future==0.16.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.6.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.10.3
jmespath==0.9.5
jsonschema==3.2.0
line-profiler==3.0.2
lxml==4.5.0
MarkupSafe==1.1.1
marshmallow==3.6.0
memory-profiler==0.57.0
more-itertools==8.2.0
murmurhash==1.0.2
newrelic==5.12.0.140
numpy==1.18.1
packaging==20.3
pandas==0.25.3
paramiko==2.7.1
parso==0.7.0
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prompt-toolkit==3.0.5
psutil==5.7.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.1
pyahocorasick==1.4.0
pycountry==19.8.18
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
redis==3.5.3
requests==2.22.0
s3transfer==0.3.3
semantic-version==2.5.0
six==1.13.0
soupsieve==2.0
spacy==2.2.3
SQLAlchemy==1.3.17
srsly==1.0.1
stripe==2.48.0
termcolor==1.1.0
texttable==1.6.2
thinc==7.3.1
tqdm==4.41.1
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.7
wasabi==0.6.0
wcwidth==0.1.9
webargs==6.1.0
websocket-client==0.57.0
Werkzeug==0.16.0
zipp==3.1.0
